refactor: previousLotto, bonusBall 합쳐서 PreviousLotto 객체 생성하도록 변경 / List<reward> 를 Game 클래스로 rollback
